# docs/swagger-inline.yaml
openapi: 3.0.0
paths:
  /auth/login:
    post:
      operationId: AuthController_login
      summary: 用户登录
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: 用户名
                  example: admin
                password:
                  type: string
                  description: 密码
                  example: '123456'
              required:
                - username
                - password
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: 访问令牌
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    description: 刷新令牌
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    description: 用户信息
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: number
                            description: 用户ID
                            example: 1
                          username:
                            type: string
                            description: 用户名
                            example: admin
                          name:
                            type: string
                            description: 姓名
                            example: 系统管理员
                          email:
                            type: string
                            description: 邮箱
                            example: admin@example.com
                          phone:
                            type: string
                            description: 电话
                            example: '13800138000'
                          avatar:
                            type: string
                            description: 头像
                            example: avatar.jpg
                          status:
                            type: number
                            description: 状态
                            example: 1
                          join_date:
                            format: date-time
                            type: string
                            description: 入职日期
                            example: '2023-01-01'
                          position:
                            type: string
                            description: 职位
                            example: 系统管理员
                          department:
                            description: 部门信息
                            allOf:
                              - type: object
                                properties:
                                  id:
                                    type: number
                                    description: 部门ID
                                    example: 1
                                  name:
                                    type: string
                                    description: 部门名称
                                    example: 技术部
                                  description:
                                    type: string
                                    description: 部门描述
                                    example: 负责技术研发工作
                                  parent_id:
                                    type: number
                                    description: 父部门ID
                                    example: 1
                                  sort_order:
                                    type: number
                                    description: 排序
                                    example: 0
                                  status:
                                    type: number
                                    description: 状态
                                    example: 1
                                  employeeCount:
                                    type: number
                                    description: 员工数量
                                    example: 10
                                  created_at:
                                    format: date-time
                                    type: string
                                    description: 创建时间
                                    example: '2023-01-01T00:00:00Z'
                                  updated_at:
                                    format: date-time
                                    type: string
                                    description: 更新时间
                                    example: '2023-01-01T00:00:00Z'
                                required:
                                  - id
                                  - name
                                  - sort_order
                                  - status
                                  - created_at
                                  - updated_at
                          leader:
                            description: 直属领导
                            allOf:
                              - type: object
                                properties:
                                  id:
                                    type: number
                                    description: 用户ID
                                    example: 1
                                  username:
                                    type: string
                                    description: 用户名
                                    example: admin
                                  name:
                                    type: string
                                    description: 姓名
                                    example: 系统管理员
                                  position:
                                    type: string
                                    description: 职位
                                    example: 系统管理员
                                required:
                                  - id
                                  - username
                                  - name
                          roles:
                            description: 角色列表
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: number
                                  description: 角色ID
                                  example: 1
                                code:
                                  type: string
                                  description: 角色代码
                                  example: admin
                                name:
                                  type: string
                                  description: 角色名称
                                  example: 系统管理员
                                description:
                                  type: string
                                  description: 角色描述
                                  example: 系统管理员角色
                                permissions:
                                  description: 权限列表
                                  example:
                                    - user:read
                                    - user:write
                                  type: array
                                  items:
                                    type: string
                                status:
                                  type: number
                                  description: 状态
                                  example: 1
                                created_at:
                                  format: date-time
                                  type: string
                                  description: 创建时间
                                  example: '2023-01-01T00:00:00Z'
                                updated_at:
                                  format: date-time
                                  type: string
                                  description: 更新时间
                                  example: '2023-01-01T00:00:00Z'
                              required:
                                - id
                                - code
                                - name
                                - status
                                - created_at
                                - updated_at
                          created_at:
                            format: date-time
                            type: string
                            description: 创建时间
                            example: '2023-01-01T00:00:00Z'
                          updated_at:
                            format: date-time
                            type: string
                            description: 更新时间
                            example: '2023-01-01T00:00:00Z'
                        required:
                          - id
                          - username
                          - name
                          - status
                          - roles
                          - created_at
                          - updated_at
                required:
                  - access_token
                  - refresh_token
                  - user
        '401':
          description: 用户名或密码错误
      tags:
        - 认证授权
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: 刷新令牌
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: 刷新令牌
              required:
                - refresh_token
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: 访问令牌
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    description: 刷新令牌
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                required:
                  - access_token
                  - refresh_token
        '401':
          description: 刷新令牌无效
      tags:
        - 认证授权
  /auth/logout:
    post:
      operationId: AuthController_logout
      summary: 用户登出
      parameters: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 消息内容
                    example: 操作成功
                required:
                  - message
        '401':
          description: 未授权
      tags:
        - 认证授权
      security:
        - bearer: []
  /auth/profile:
    get:
      operationId: AuthController_getProfile
      summary: 获取当前用户信息
      parameters: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: 用户ID
                    example: 1
                  username:
                    type: string
                    description: 用户名
                    example: admin
                  name:
                    type: string
                    description: 姓名
                    example: 系统管理员
                  email:
                    type: string
                    description: 邮箱
                    example: admin@example.com
                  phone:
                    type: string
                    description: 电话
                    example: '13800138000'
                  avatar:
                    type: string
                    description: 头像
                    example: avatar.jpg
                  status:
                    type: number
                    description: 状态
                    example: 1
                  join_date:
                    format: date-time
                    type: string
                    description: 入职日期
                    example: '2023-01-01'
                  position:
                    type: string
                    description: 职位
                    example: 系统管理员
                  department:
                    description: 部门信息
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: number
                            description: 部门ID
                            example: 1
                          name:
                            type: string
                            description: 部门名称
                            example: 技术部
                          description:
                            type: string
                            description: 部门描述
                            example: 负责技术研发工作
                          parent_id:
                            type: number
                            description: 父部门ID
                            example: 1
                          sort_order:
                            type: number
                            description: 排序
                            example: 0
                          status:
                            type: number
                            description: 状态
                            example: 1
                          employeeCount:
                            type: number
                            description: 员工数量
                            example: 10
                          created_at:
                            format: date-time
                            type: string
                            description: 创建时间
                            example: '2023-01-01T00:00:00Z'
                          updated_at:
                            format: date-time
                            type: string
                            description: 更新时间
                            example: '2023-01-01T00:00:00Z'
                        required:
                          - id
                          - name
                          - sort_order
                          - status
                          - created_at
                          - updated_at
                  leader:
                    description: 直属领导
                    allOf:
                      - type: object
                        properties:
                          id:
                            type: number
                            description: 用户ID
                            example: 1
                          username:
                            type: string
                            description: 用户名
                            example: admin
                          name:
                            type: string
                            description: 姓名
                            example: 系统管理员
                          position:
                            type: string
                            description: 职位
                            example: 系统管理员
                        required:
                          - id
                          - username
                          - name
                  roles:
                    description: 角色列表
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: 角色ID
                          example: 1
                        code:
                          type: string
                          description: 角色代码
                          example: admin
                        name:
                          type: string
                          description: 角色名称
                          example: 系统管理员
                        description:
                          type: string
                          description: 角色描述
                          example: 系统管理员角色
                        permissions:
                          description: 权限列表
                          example:
                            - user:read
                            - user:write
                          type: array
                          items:
                            type: string
                        status:
                          type: number
                          description: 状态
                          example: 1
                        created_at:
                          format: date-time
                          type: string
                          description: 创建时间
                          example: '2023-01-01T00:00:00Z'
                        updated_at:
                          format: date-time
                          type: string
                          description: 更新时间
                          example: '2023-01-01T00:00:00Z'
                      required:
                        - id
                        - code
                        - name
                        - status
                        - created_at
                        - updated_at
                  created_at:
                    format: date-time
                    type: string
                    description: 创建时间
                    example: '2023-01-01T00:00:00Z'
                  updated_at:
                    format: date-time
                    type: string
                    description: 更新时间
                    example: '2023-01-01T00:00:00Z'
                required:
                  - id
                  - username
                  - name
                  - status
                  - roles
                  - created_at
                  - updated_at
        '401':
          description: 未授权
      tags:
        - 认证授权
      security:
        - bearer: []
  /auth/password:
    post:
      operationId: AuthController_changePassword
      summary: 修改密码
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  description: 原密码
                newPassword:
                  type: string
                  description: 新密码
              required:
                - oldPassword
                - newPassword
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: 消息内容
                    example: 操作成功
                required:
                  - message
        '400':
          description: 原密码错误
        '401':
          description: 未授权
      tags:
        - 认证授权
      security:
        - bearer: []
...
info:
  title: OKR绩效考核系统API
  description: OKR绩效考核系统后端API文档
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...